components/backend/requirements.txt
# =============================================================================
# Python Dependencies for OmniVoIP Backend
# =============================================================================

# Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2

# Database
psycopg[binary]==3.1.18
redis==5.0.1

# Task Queue
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Authentication & Security
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0
django-allauth==0.58.2
python-decouple==3.8

# Object Storage
boto3==1.34.10
django-storages==1.14.2

# API Documentation
drf-spectacular==0.27.0
drf-spectacular-sidecar==2023.12.1

# HTTP Client
requests==2.31.0
httpx==0.25.2

# Data Validation
pydantic==2.5.3
marshmallow==3.20.1

# Date & Time
python-dateutil==2.8.2
pytz==2023.3.post1

# Utils
python-dotenv==1.0.0
python-slugify==8.0.1
Pillow==10.1.0

# Monitoring & Logging
sentry-sdk==1.39.1
python-json-logger==2.0.7

# Excel/CSV
openpyxl==3.1.2
xlsxwriter==3.1.9
pandas==2.1.4

# VoIP/Telephony
panoramisk==1.4  # Asterisk AMI
aiortc==1.6.0    # WebRTC

# Server
gunicorn==21.2.0
uvicorn[standard]==0.25.0
whitenoise==6.6.0

# Development
ipython==8.18.1
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==21.0.0
